rfb-client: Add serial port timeout
`serialport::new()` currently initializes the timeout with zero. On
Windows, this results in blocking operations,
while on POSIX systems, calls return immediately
resulting in no data being received by the read calls.
As `linux_embedded_hal::serial_unix` defaults to a 100 ms timeout, this
seems to be a reasonable timeout value to choose.